chore(cleanup): idiomatic changes
